refactor(ai): de-part Slice A planner/perception clusters (Refs #4079)
Convert perception, diplomatic scoring, conquest, recruitment, and economy
part graphs into explicit-import libraries without changing AI semantics.
